coproc: remove dependency on rpc test services
Coproc uses the rpc fixture which is generic, but the the fixture also
included the generated services used in the rpc tests. Split these
in two so that coproc could use the generic fixture without relying
on the code generated services used only by rpc internal tests.
