+++ src/v/rpc/test/rpc_integration_fixture.h
@@ -15,9 +15,6 @@
 #include "net/server.h"
 #include "rpc/service.h"
 #include "rpc/simple_protocol.h"
-#include "rpc/test/cycling_service.h"
-#include "rpc/test/echo_service.h"
-#include "rpc/test/rpc_gen_types.h"
 #include "rpc/transport.h"
 #include "rpc/types.h"
 #include "seastarx.h"
@@ -27,70 +24,6 @@
 #include <seastar/core/smp.hh>
 #include <seastar/net/inet_address.hh>
 #include <seastar/net/tls.hh>
-
-// Test services
-struct movistar final : cycling::team_movistar_service {
-    movistar(ss::scheduling_group& sc, ss::smp_service_group& ssg)
-      : cycling::team_movistar_service(sc, ssg) {}
-    ss::future<cycling::mount_tamalpais>
-    ibis_hakka(cycling::san_francisco&&, rpc::streaming_context&) final {
-        return ss::make_ready_future<cycling::mount_tamalpais>(
-          cycling::mount_tamalpais{66});
-    }
-    ss::future<cycling::nairo_quintana>
-    canyon(cycling::ultimate_cf_slx&&, rpc::streaming_context&) final {
-        return ss::make_ready_future<cycling::nairo_quintana>(
-          cycling::nairo_quintana{32});
-    }
-};
-
-struct echo_impl final : echo::echo_service {
-    echo_impl(ss::scheduling_group& sc, ss::smp_service_group& ssg)
-      : echo::echo_service(sc, ssg) {}
-    ss::future<echo::echo_resp>
-    echo(echo::echo_req&& req, rpc::streaming_context&) final {
-        return ss::make_ready_future<echo::echo_resp>(
-          echo::echo_resp{.str = req.str});
-    }
-    ss::future<echo::echo_resp>
-    prefix_echo(echo::echo_req&& req, rpc::streaming_context&) final {
-        return ss::make_ready_future<echo::echo_resp>(
-          echo::echo_resp{.str = ssx::sformat("prefix_{}", req.str)});
-    }
-    ss::future<echo::echo_resp>
-    suffix_echo(echo::echo_req&& req, rpc::streaming_context&) final {
-        return ss::make_ready_future<echo::echo_resp>(
-          echo::echo_resp{.str = ssx::sformat("{}_suffix", req.str)});
-    }
-
-    ss::future<echo::echo_resp>
-    sleep_1s(echo::echo_req&&, rpc::streaming_context&) final {
-        using namespace std::chrono_literals;
-        return ss::sleep(1s).then(
-          []() { return echo::echo_resp{.str = "Zzz..."}; });
-    }
-
-    ss::future<echo::cnt_resp>
-    counter(echo::cnt_req&& req, rpc::streaming_context&) final {
-        return ss::make_ready_future<echo::cnt_resp>(
-          echo::cnt_resp{.expected = req.expected, .current = cnt++});
-    }
-
-    ss::future<echo::throw_resp>
-    throw_exception(echo::throw_req&& req, rpc::streaming_context&) final {
-        switch (req) {
-        case echo::failure_type::exceptional_future:
-            return ss::make_exception_future<echo::throw_resp>(
-              std::runtime_error("gentle crash"));
-        case echo::failure_type::throw_exception:
-            throw std::runtime_error("bad crash");
-        default:
-            return ss::make_ready_future<echo::throw_resp>();
-        }
-    }
-
-    uint64_t cnt = 0;
-};
 
 class rpc_base_integration_fixture {
 public:
@@ -252,17 +185,4 @@ private:
     }
 
     ss::sharded<net::server> _server;
-};
-
-class rpc_integration_fixture : public rpc_simple_integration_fixture {
-public:
-    rpc_integration_fixture()
-      : rpc_simple_integration_fixture(redpanda_rpc_port) {}
-
-    void register_services() {
-        register_service<movistar>();
-        register_service<echo_impl>();
-    }
-
-    static constexpr uint16_t redpanda_rpc_port = 32147;
 };
